feat(devtools): add values and description columns to props table
Enhance the `ComponentsBookPage.vue` by including `Values` and
`Description` columns to the props table. Update the `extrector.ts`
to parse and return these values. Modify the `nuxt.config.ts` to
disable caching for the components book. Update the version in
`package.json` to `1.0.6`. Adjust the sidebar width in
`ComponentsBookLayout.vue` for better layout.